+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Polyverse Boost Automatic Application Modernization
 - Fix issue with Test and File generation markdown rendering of plain-text
 - Fix issue with Test Generation for some historic languages (COBOL) that don't have dedicated Test Frameworks failing to generate any Test Cases
 - Fix issue with incorrect language being reported in PDF, HTML and Markdown output
+- Fix issue where details of analysis/code generation are not stored in Boost files when using the Notebook UI
 
 ## Version 1.6.4: November 3rd, 2023